Fix task link shortcuts where space in target path
Addresses #19 where if if the true target path contains spaces, the script was incorrectly was splitting target/args at the wrong place therefore creating invalid shortcuts.

Note: In the run box the target string will appear the same but behind the scenes it will be fixed because of how shortcuts are handled.